collect tests by method
